Rename "lossless storage" to "dict storage"
