+++ camel/memory/chat_history_memory.py
@@ -17,8 +17,8 @@ from dataclasses import asdict
 from typing import List, Optional
 
 from camel.memory.base_memory import BaseMemory
-from camel.memory.lossless_storage.base import LosslessStorage
-from camel.memory.lossless_storage.in_memory import InMemoryStorage
+from camel.memory.dict_storage.base import DictStorage
+from camel.memory.dict_storage.in_memory import InMemoryDictStorage
 from camel.messages.base import BaseMessage
 
 
@@ -38,16 +38,16 @@ class ChatHistoryMemory(BaseMemory):
     the chat history.
 
     Args:
-        storage (LosslessStorage): A storage mechanism for storing chat
+        storage (DictStorage): A storage mechanism for storing chat
             history.
         window_size (int, optional): Specifies the number of recent chat
             messages to retrieve. If not provided, the entire chat history
             will be retrieved.
     """
 
-    def __init__(self, storage: Optional[LosslessStorage] = None,
+    def __init__(self, storage: Optional[DictStorage] = None,
                  window_size: Optional[int] = None) -> None:
-        self.storage = storage or InMemoryStorage()
+        self.storage = storage or InMemoryDictStorage()
         self.window_size = window_size
 
     def read(self,
