+++ camel/memory/dict_storage/in_memory.py
@@ -12,12 +12,27 @@
 # limitations under the License.
 # =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
 
-from .base import LosslessStorage
-from .in_memory import InMemoryStorage
-from .json import JsonStorage
+from copy import deepcopy
+from typing import Any, Dict, List
 
-__all__ = [
-    'LosslessStorage',
-    'InMemoryStorage',
-    'JsonStorage',
-]
+from camel.memory.dict_storage.base import DictStorage
+
+
+class InMemoryDictStorage(DictStorage):
+    """
+    Concrete implementation of the :obj:`DictStorage` using in-memory list.
+    Ideal for temporary storage purposes, as data will be lost when the program
+    ends.
+    """
+
+    def __init__(self) -> None:
+        self.memory_list: List[Dict] = []
+
+    def save(self, records: List[Dict[str, Any]]) -> None:
+        self.memory_list.extend(deepcopy(records))
+
+    def load(self) -> List[Dict[str, Any]]:
+        return deepcopy(self.memory_list)
+
+    def clear(self) -> None:
+        self.memory_list.clear()
