+++ docs/camel.memory.rst
@@ -8,7 +8,7 @@ Subpackages
    :maxdepth: 4
    
    camel.memory.vector_storage
-   camel.memory.lossless_storage
+   camel.memory.dict_storage
 
 
 Module contents
